Finalized Rating system
Added function to the rating buttons. They now add to an array of numbers
to represent all ratings.
Rating is displayed on the top, near the name of the milk.
Rating updates as it gets rated, displaying on the top.
When there are no ratings (aka, rating is 0), message is displayed
letting the user know there are no ratings yet.