Added admin function to clear ratings
Added button to allow the admin to clear the ratings from its component
page.
Clears the array of ratings and sets calcRating to 0 when clicked.
Made modifications to milk.component.html to the passed number
Made modifications to milk-detail.component.html to organize the fields.
Added JavaDoc comments to the methods in Milk.java